+++ src/main/java/com/devone/bot/core/utils/blocks/BotPositionComparators.java
@@ -6,37 +6,37 @@ public class BotPositionComparators {
 
     public static final Comparator<BotPosition> X_ASC = new Comparator<>() {
         @Override public int compare(BotPosition a, BotPosition b) {
-            return Integer.compare(a.getX(), b.getX());
+            return Double.compare(a.getX(), b.getX());
         }
     };
 
     public static final Comparator<BotPosition> X_DESC = new Comparator<>() {
         @Override public int compare(BotPosition a, BotPosition b) {
-            return Integer.compare(b.getX(), a.getX());
+            return Double.compare(b.getX(), a.getX());
         }
     };
 
     public static final Comparator<BotPosition> Y_ASC = new Comparator<>() {
         @Override public int compare(BotPosition a, BotPosition b) {
-            return Integer.compare(a.getY(), b.getY());
+            return Double.compare(a.getY(), b.getY());
         }
     };
 
     public static final Comparator<BotPosition> Y_DESC = new Comparator<>() {
         @Override public int compare(BotPosition a, BotPosition b) {
-            return Integer.compare(b.getY(), a.getY());
+            return Double.compare(b.getY(), a.getY());
         }
     };
 
     public static final Comparator<BotPosition> Z_ASC = new Comparator<>() {
         @Override public int compare(BotPosition a, BotPosition b) {
-            return Integer.compare(a.getZ(), b.getZ());
+            return Double.compare(a.getZ(), b.getZ());
         }
     };
 
     public static final Comparator<BotPosition> Z_DESC = new Comparator<>() {
         @Override public int compare(BotPosition a, BotPosition b) {
-            return Integer.compare(b.getZ(), a.getZ());
+            return Double.compare(b.getZ(), a.getZ());
         }
     };
